refactor: folder of components
