+++ src/components/products/ProductBox/ProductDetails.tsx
