+++ src/components/steps/StepBar.tsx
