+++ src/pages/Cart.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
-import ProductList from '@/components/ProductList'
-import ResumeValues from '@/components/ResumeOrder'
+import ProductList from '@/components/products/ProductList'
+import ResumeValues from '@/components/resume/ResumeValues'
 
 function Cart (): React.ReactElement | null {
   return (
